api/requirements.txt
alabaster==0.7.16
babel==2.17.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.1.8
db-dtypes==1.4.3
docutils==0.21.2
Flask==3.1.2
flask-cors==6.0.1
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.2
google-cloud-bigquery==3.38.0
google-cloud-bigquery-connection==1.18.3
google-cloud-bigquery-storage==2.33.1
google-cloud-core==2.4.3
google-cloud-storage==2.19.0
google-crc32c==1.7.1
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
grpc-google-iam-v1==0.14.2
grpcio==1.75.0
grpcio-status==1.75.0
httplib2==0.31.0
idna==3.10
imagesize==1.4.1
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
load-dotenv==0.1.0
loguru==0.7.3
MarkupSafe==3.0.2
numpy==2.0.2
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pandas-gbq==0.29.2
proto-plus==1.26.1
protobuf==6.32.1
pyarrow==21.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydata-google-auth==1.9.1
Pygments==2.19.2
pymssql==2.3.7
pyodbc==5.2.0
pyparsing==3.2.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
six==1.17.0
snowballstemmer==3.0.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
SQLAlchemy==2.0.43
tomli==2.2.1
tomlkit==0.11.8
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
Werkzeug==3.1.3
zipp==3.23.0
